Labels, Rooms, File reading
Added labels, text that is drawn to the panel. The text wraps.
Additionally labels are positioned dynamically, in relation to other
labels or their "zone," a rectangular area of the panel. Additionally,
basic file reading has been added, and room data can now be loaded and
displayed, though nothing is yet done with that data.